do not pass socket
