+++ frontend/src/App.tsx
@@ -1,15 +1,38 @@
+import { useState, useEffect } from 'react';
+
 import { io } from 'socket.io-client';
 
+import { RegType } from './features/regs/types/RegType';
 import { RegsArea } from './components/RegsArea';
 
 const socket = io('http://localhost:8000');
 
 function App() {
+  const [regsArr, setRegsArr] = useState<RegType[]>([]);
+
   socket.on('connect', () => {
     console.log('socket connect', socket.connect());
   });
 
-  return <RegsArea socket={socket} />;
+  useEffect(() => {
+    socket.on('get_regs', (res) => {
+      setRegsArr(res);
+    });
+  }, []);
+
+  const clickAct = () => {
+    const res = socket.emit('get_regs');
+    console.log('CLICK ACT', res);
+  };
+
+  return (
+    <div>
+      <h1>DE</h1>
+      <hr />
+      <button onClick={clickAct}>Get Info</button>
+      <RegsArea regsArr={regsArr} />
+    </div>
+  );
 }
 
 export default App;
