+++ frontend/src/components/RegsArea.tsx
@@ -1,27 +1,10 @@
-import { Socket } from 'socket.io-client';
-import { useEffect, useState } from 'react';
-
 import { RegType } from '../features/regs/types/RegType';
 import { ShowRegs } from '../features/regs/components/ShowRegs';
 
-export const RegsArea = ({ socket }: { socket: Socket }) => {
-  const [regsArr, setRegsArr] = useState<RegType[]>([]);
-
-  useEffect(() => {
-    socket.on('get_regs', (res) => {
-      setRegsArr(res);
-    });
-  }, []);
-
-  const clickAct = () => {
-    const res = socket.emit('get_regs');
-    console.log('CLICK ACT', res);
-  };
-
+export const RegsArea = ({ regsArr }: { regsArr: RegType[] }) => {
   return (
     <div>
       <h1>RegsArea</h1>
-      <button onClick={clickAct}>Get Regs</button>
       <ShowRegs regs={regsArr} />
     </div>
   );
